api/requirements.txt
Flask-Cors~=3.0.6
Flask-HTTPAuth~=3.2.4
Flask-RESTful~=0.3.6
gunicorn~=19.9.0
pymongo
